web: move non-deprecated utilities to sopel.tools.web
The deprecated functions remain available only from `sopel.web`, to be
left behind as plugin authors transition their code to the new import
location. Nothing imported from `sopel.web` should break.

In the new `tools.web` module, both `USER_AGENT` and `DEFAULT_HEADERS`
should be all-caps, so they really do look like constants.

`web.ca_certs` appears to have been irrelevant for ages now. Left it in
the old import location just in case some silly module is using it for
something (don't know what; it'll always be `None` now), and it will go
away when `sopel.web` is deleted completely in Sopel 8.0..

Removed imports of `ssl` module and all the gymnastics about getting
`match_hostname` from backports if needed, since nothing in `web` calls
on the `ssl` module directly any more. Not even the wrapper functions.
